Переносит всю логику по отрисовке фильмов в контроллер
